move setup components into setup folder
