+++ src/pages/Setup/FormatDialog.tsx
@@ -4,7 +4,7 @@ import { ReactComponent as DBIcon } from "../../assets/database.svg";
 import { ReactComponent as TrashIcon } from "../../assets/trash.svg";
 import { ReactComponent as X } from "../../assets/X.svg";
 import SetupContainer from "../../container/SetupContainer/SetupContainer";
-import ConfirmModal from "../Shared/ConfirmModal/ConfirmModal";
+import ConfirmModal from "../../components/Shared/ConfirmModal/ConfirmModal";
 
 export type Props = {
   containsBlockchain: boolean;
